+++ doc/source/reference/routines.io.rst
@@ -81,7 +81,7 @@ Data sources
 .. autosummary::
    :toctree: generated/
 
-   DataSource
+   lib.npyio.DataSource
 
 Binary format description
 -------------------------
@@ -89,3 +89,10 @@ Binary format description
    :toctree: generated/
 
    lib.format
+
+Object wrappers
+---------------
+.. autosummary::
+   :toctree: generated/
+
+   lib.npyio.BagObj
