+++ numpy/lib/npyio.pyi
@@ -1,4 +1,5 @@
 from numpy.lib._npyio_impl import (
     BagObj as BagObj,
+    DataSource as DataSource,
     NpzFile as NpzFile,
 )
