+++ numpy/typing/tests/data/reveal/datasource.pyi
@@ -4,9 +4,9 @@ import numpy as np
 path1: Path
 path2: str
 
-d1 = np.DataSource(path1)
-d2 = np.DataSource(path2)
-d3 = np.DataSource(None)
+d1 = np.lib.npyio.DataSource(path1)
+d2 = np.lib.npyio.DataSource(path2)
+d3 = np.lib.npyio.DataSource(None)
 
 reveal_type(d1.abspath("..."))  # E: str
 reveal_type(d2.abspath("..."))  # E: str
